Moved parallel_split_filter outside get_composite_source_model [skip hazardlib][demos]
